move GetApplicationOrError from application helper to service
