+++ src/Storage/Services/ApplicationService.cs
@@ -1,8 +1,9 @@
-using System.Linq;
+using System;
 using System.Threading.Tasks;
 using Altinn.Platform.Storage.Interface.Models;
 using Altinn.Platform.Storage.Models;
 using Altinn.Platform.Storage.Repository;
+using Microsoft.AspNetCore.Mvc;
 
 namespace Altinn.Platform.Storage.Services
 {
@@ -12,7 +13,7 @@ namespace Altinn.Platform.Storage.Services
     public class ApplicationService : IApplicationService
     {
         private readonly IApplicationRepository _applicationRepository;
-        
+
         /// <summary>
         /// Initializes a new instance of the <see cref="ApplicationService"/> class.
         /// </summary>
@@ -20,7 +21,7 @@ namespace Altinn.Platform.Storage.Services
         {
             _applicationRepository = applicationRepository;
         }
-        
+
         /// <inheritdoc/>
         public async Task<(bool IsValid, ServiceError ServiceError)> ValidateDataTypeForApp(string org, string appId, string dataType, string currentTask)
         {
@@ -37,6 +38,38 @@ namespace Altinn.Platform.Storage.Services
             }
 
             return (false, new ServiceError(405, $"DataType {dataType} is not declared in application metadata for app {appId}"));
+        }
+
+        /// <summary>
+        /// Wrapper method for getting an application or an error message using the application id.
+        /// </summary>
+        /// <param name="appId">The application id</param>
+        /// <returns></returns>
+        public async Task<(Application Application, ActionResult ErrorMessage)> GetApplicationOrErrorAsync(string appId)
+        {
+            ActionResult errorResult = null;
+            Application appInfo = null;
+
+            try
+            {
+                string org = appId.Split("/")[0];
+
+                appInfo = await _applicationRepository.FindOne(appId, org);
+
+                if (appInfo == null)
+                {
+                    errorResult = new NotFoundObjectResult($"Did not find application with appId={appId}");
+                }
+            }
+            catch (Exception e)
+            {
+                errorResult = new ObjectResult($"Unable to perform request: {e}")
+                {
+                    StatusCode = 500
+                };
+            }
+
+            return (appInfo, errorResult);
         }
     }
 }
