+++ src/Storage/Services/IApplicationService.cs
@@ -1,5 +1,7 @@
 using System.Threading.Tasks;
+using Altinn.Platform.Storage.Interface.Models;
 using Altinn.Platform.Storage.Models;
+using Microsoft.AspNetCore.Mvc;
 
 namespace Altinn.Platform.Storage.Services
 {
@@ -16,6 +18,13 @@ namespace Altinn.Platform.Storage.Services
         /// <param name="dataType">The data type identifier for the data being uploaded.</param>
         /// <param name="currentTask">The task info of the currentTask of an ongoing process.</param>
         /// <returns>Result of validation. If the result (IsValid) is false, it will be described in ServiceError</returns>
-         Task<(bool IsValid, ServiceError ServiceError)> ValidateDataTypeForApp(string org, string appId, string dataType, string currentTask);
+        Task<(bool IsValid, ServiceError ServiceError)> ValidateDataTypeForApp(string org, string appId, string dataType, string currentTask);
+
+        /// <summary>
+        /// Get application or error message using the application id.
+        /// </summary>
+        /// <param name="appId">The id of the application.</param>
+        /// <returns>A <see cref="Task{TResult}"/> representing the result of the asynchronous operation.</returns>
+        Task<(Application Application, ActionResult ErrorMessage)> GetApplicationOrErrorAsync(string appId);
     }
 }
